backend/requirements-dev.txt
# Type checking
mypy>=1.0.0
types-boto3>=1.0.0
mypy-boto3-dynamodb>=1.26.0

# Development
black>=23.0.0
isort>=5.0.0
flake8>=6.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0 